// src/models/interfaces/photo.schema.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * A valid photo object, with license info.
 */
export type PhotoSchema = {
  [k: string]: unknown;
} & {
  /**
   * ID of the image to embed via SmartFrame.io
   */
  embedId?: number;
  /**
   * The URL of the photo.
   */
  url?: string;
  /**
   * Description of the picture.
   */
  description?: string;
  /**
   * Author name of the photo.
   */
  author?: string;
  /**
   * Web site name of the URL provided.
   */
  site?: string;
  /**
   * Web site where the photo is. Not to be confused with the URL of the photo itself. In this, in addition to the photo, there is usually other information such as the author, date of capture...
   */
  siteURL?: string;
  /**
   * The CC license type.
   */
  CCType?: 'BY' | 'BY-SA' | 'BY-ND' | 'BY-NC' | 'BY-NC-ND' | 'Public Domain';
  /**
   * The CC license version number (if not CC public domain)
   */
  CC?: number;
};